Fix some warnings

+++ src/dbg/commands/cmd-undocumented.cpp
@@ -529,7 +529,7 @@ bool cbInstrDebugFlags(int argc, char* argv[])
         dprintf_untranslated("Usage: DebugFlags 0xFFFFFFFF\n");
         return false;
     }
-    auto debugFlags = DbgValFromString(argv[1]);
+    auto debugFlags = (DWORD)DbgValFromString(argv[1]);
     dbgsetdebugflags(debugFlags);
     dprintf_untranslated("DebugFlags = 0x%08X\n", debugFlags);
     return true;
